# OCR-Spring-Backend-main/swagger/ocr-api.yaml
openapi: 3.0.0
info:
  title: OCR API
  description: API for managing OCR apps
  version: 1.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://localhost
...
paths:
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Sign up a new user
      description: >
        This endpoint adds a new user signup request. If the registration number or email is already registered,
        it returns an appropriate message. If a signup request already exists for the registration number or email,
        it returns a message indicating that. Otherwise, it creates a new signup request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SignupSuccess"
                  - $ref: "#/components/schemas/SignupExists"
        "401":
          description: Registration number or email is already registered
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RegNoError"
                  - $ref: "#/components/schemas/EmailError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify a user
      description: >
        This endpoint verifies a user signup request. If the request is valid, it sends an email to the user.
        If the request is not found, it returns an error message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"

      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginSuccess"

        "400":
          description: User is not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotRegisteredError"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/refreshToken:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: >
        This endpoint refreshes the access token using a provided refresh token.
      responses:
        "200":
          description: >
            Refresh token successful.
            Note: Although a 200 status code is not explicitly set in the code, this response is implied for successful refresh token operations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenSuccess"
        "400":
          description: Token is required or an error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenError"
  /api/auth/revokeToken:
    post:
      tags:
        - Authentication
      summary: Revoke a token
      description: >
        This endpoint revokes an access or refresh token. The token can be provided in the request body or in cookies.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeTokenRequest"
      responses:
        "200":
          description: >
            Token revoked successfully.
            Note: Although a 200 status code is not explicitly set in the code, this response is implied for successful revocation operations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeTokenSuccess"
        "400":
          description: Token is required or an error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeTokenError"
        "403":
          description: User is not authorized to revoke this token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
...
  /api/admin/auth/signup:
    post:
      tags: 
        - Admin Authentication
      summary: To add the initial admin
      description: >
        This endpoint adds the initial admin to the system. If the admin already exists, it returns an error message.
      requestBody:
        required: true
        content:
          application/json:
            schema:  
              $ref: "#/components/schemas/NewAdminSignUp"
      responses: 
        "200":
          description: "Successfully signed in"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewAdminSignUpResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error!"
                  error:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /api/admin/update/user/{id}:
    post:
      summary: update user details
      tags:
        - Admin
      description: >
        This endpoint updates the details of a user. If the user is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/details"
      responses:
        "200":
          description: successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/details"
        
        "401":
          description: unauthorized access
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/UnauthorizedResponse"

        "404":
          description: user not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/RequestNotFoundResponse"

        "500": 
          description: internal server error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /api/admin/delete/user/{id}:
    post:
      summary: delete user
      tags:
        - Admin
      description: > 
        This endpoint deletes a user. If the user is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to delete
      responses:
        "200":
          description: successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userdelete"
        
        "401":
          description: unauthorized access
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/UnauthorizedResponse"

        "404":
          description: user not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/RequestNotFoundResponse"

        "500": 
          description: internal server error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/InternalServerErrorResponse"
...
  /api/user/self:
    get:
      summary: Get a user
      tags:
        - User
      description: >
        This endpoint returns the details of the authenticated user.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
  /api/user/self/hospitals:
    get:
      description: >
        This endpoint returns the list of hospitals available for the user to select.
      summary: Get hospital list on signup page
      tags:
        - User
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/hospital"
        "500":
          description: Internal Server Error
  /api/user/self/update:
    post:
      summary: Update a user
      tags:
        - User
      description: >
        This endpoint allows users to update their details. If the user is not found, it returns an error message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User details updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error            
...
components:
  schemas:
    User:
      type: object
      properties:
        username: 
          type: string
        email:
          type: string
        reg_no:
          type: string
        hospital:
          type: string
        designation:
          type: string
        contact_no:
          type: string
        availability:
          type: string
        role:
          type: string
    hospital:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        city:
          type: string
        address:
          type: string
        contact_no:
          type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        hospital:
          type: string
        contact_no:
          type: string
        availability:
          type: boolean
    LoginRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          description: Email address
          example: johndoe@example.com
        password:
          type: string
          description: Password
          example: "yourpassword"
    SignupRequest:
      type: object
      required:
        - reg_no
        - username
        - email
      properties:
        reg_no:
          type: string
          description: Registration number
          example: ABC123
        username:
          type: string
          description: Username
          example: johndoe
        email:
          type: string
          description: Email address
          example: johndoe@example.com
        hospital:
          type: string
          description: Name of the hospital
          example: General Hospital
        designation:
          type: string
          description: Designation of the user
          example: Doctor
        contact_no:
          type: string
          description: Contact number
          example: "+1234567890"
    SignupSuccess:
      type: object
      properties:
        username:
          type: string
          description: Username
        email:
          type: string
          description: Email address
        reg_no:
          type: string
          description: Registration number
        hospital:
          type: string
          description: Name of the hospital
        designation:
          type: string
          description: Designation of the user
        contact_no:
          type: string
          description: Contact number
        _id:
          type: string
          description: User ID
        createdAt:
          type: string
          description: Date of creation
        updatedAt:
          type: string
          description: Date of last update
        message:
          type: string
          description: Success message
          example: "Request is sent successfully. You will receive an Email on acceptance."
    SignupExists:
      type: object
      properties:
        message:
          type: string
          description: Message indicating an existing request
          example: "A request for registration already exists."
    RegNoError:
      type: object
      properties:
        message:
          type: string
          description: Error message for registered registration number
          example: "The Reg No is already registered"
    EmailError:
      type: object
      properties:
        message:
          type: string
          description: Error message for registered email address
          example: "Email address is already registered"
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Internal Server Error!"
        error:
          type: string
          description: Detailed error information
          example: "Detailed error information here."
    userref:
          type: object
          description: User reference details
          properties:
            _id:
              type: string
              description: User ID
              example: "60b6b3b8f9c0c40015c10f5a"
            username:
              type: string
              description: Username
              example: "johndoe"
            email:
              type: string
              description: User's email address
              example: "johndoe@example.com"
            reg_no:
              type: string
              description: Registration number
              example: "1"
            hospital:
              type: string
              description: Name of the hospital
              example: "Colombo Hospital"
            designation:
              type: string
              description: Designation of the user
              example: "Doctor"
            contact_no:
              type: string
              description: Contact number
              example: "+1234567890"
            password:
              type: string
              description: Encrypted password
              example: "$2a$10$3Q"
            availability:
              type: boolean
              description: Availability status
              example: true
            role:
              type: string
              description: User's role
              example: "admin"
            createdAt:
              type: string
              description: Date of creation
              example: "2021-06-01T12:00:00.000Z"
            updatedAt:
              type: string
              description: Date of last update
              example: "2021-06-01T12:00:00.000Z" 
            message:
              type: string
              description: Success message
              example: "Successfully logged in"
            permissions: 
              type: array
              example: [300]
    LoginSuccess:
      type: object
      properties:
        accessToken:
          type: object
          properties:
            token:
              type: string
              description: Access token
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            expiry:
              type: string
              description: Token expiry time
              example: "1h"
        ref:
          description: User reference details
          type: object
          allOf:
            - $ref: "#/components/schemas/userref"
        others:
          description: User reference details
          type: object
          allOf:
            - $ref: "#/components/schemas/userref"
    UserNotRegisteredError:
      type: object
      properties:
        message:
          type: string
          description: Error message indicating the user is not registered
          example: "User is not registered!"
    RefreshTokenSuccess:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        message:
          type: string
          description: Success message
          example: "Refresh token successful"
        ref:
          type: object
          description: User reference details
          properties:
            _id:
              type: string
              example: "60b6b3b8f9c0c40015c10f5a"
            username:
              type: string
              example: "johndoe"
            email:
              type: string
              example: "johndoe@example.com"
            reg_no:
              type: string
              example: "1"
            hospital:
              type: string
              example: "Colombo Hospital"
            designation:
              type: string
              example: "Doctor"
            contact_no:
              type: string
              example: "+1234567890"
            password:
              type: string
              example: "$2a$10$3Q"
            availability:
              type: boolean
              example: true
            role:
              type: string
              example: "admin"
            createdAt:
              type: string
              example: "2021-06-01T12:00:00.000Z"
            updatedAt:
              type: string
              example: "2021-06-01T12:00:00.000Z" 
        permissions:
          type: array
          items:
            type: integer
          description: List of user permissions
          example: [300]
        accessToken:
          type: object
          properties:
            token:
              type: string
              description: Access token
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            expiry:
              type: string
              description: Token expiry time
              example: "1h"
    RefreshTokenError:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: false
        message:
          type: string
          description: Error message
          example: "Token is required"
        error:
          type: string
          description: Detailed error information
          example: "Detailed error information here."
    RevokeTokenRequest:
      type: object
      properties:
        accessToken:
          type: string
          description: Access token to revoke
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RevokeTokenSuccess:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        message:
          type: string
          description: Success message
          example: "Token revoked"
    RevokeTokenError:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: false
        message:
          type: string
          description: Error message
          example: "Token is required"
    UnauthorizedError:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: false
        message:
          type: string
          description: Error message
          example: "You are not authorized to revoke this token"
...
    NewAdminSignUp:
      type: object
      properties:
        reg_no:  
          type: integer
          description: clinician reg_no
          example: 96905
        username: 
          type: string
          description: clinician name
          example: "Nethmini Karunarathne"
        email: 
          type: string
          description: clinician email
          example: "e20189@eng.pdn.ac.lk"
        hospital: 
          type: string
          description: clinician hospital
          example: "Peradeniya Teaching Hospital"
    NewAdminSignUpResponse:
      type: object
      properties:
        username: 
          type: string
          example: "Nethmini Karunarathne"
        email: 
          type: string
          example: "e20189@eng.pdn.ac.lk"
        reg_no: 
          type: string
          example: "96964"
        hospital: 
          type: string
        designation: 
          type: string
          example: ""
        contact_no: 
          type: string
          example: ""
        availability: 
          type: boolean
          example: true
        role: 
          type: string
          example: "System Admin"
        _id: 
          type: string
          example: "60b6b3b8f9c0c40015c10f5a"
        createdAt: 
          type: string
          example: "2024-07-11T01:35:29.190Z"
        updatedAt: 
          type: string
          example: "2024-07-11T01:35:29.190Z" 
        message: 
          type: string
          example: "Successfully signed in"
...
    details:
      type: object
      properties:
        username: 
          type: string
        email:
          type: string
        reg_no:
          type: string
        hospital:
          type: string
        designation:
          type: string
        contact_no:
          type: string
        availability:
          type: string
        role:
          type: string
        message:
          type: string
          default: "user details updated successfully"
    userdelete:
      type: object
      properties:
        message:
          type: string
          default: "user deleted successfully"
...
    UnauthorizedResponse:
      type: object
      properties:
        message:
          type: string
          default: "Unauthorized access"
...
    RequestNotFoundResponse:
      type: object
      properties:
        message:
          type: string
          default: "Request not found"
...
    InternalServerErrorResponse:
      type: object
      properties:
        message:
          type: string
          default: "Internal server error"
...